requirements.txt
opencv-contrib-python
imutils==0.5.4
matplotlib==3.3.3
numpy==1.18.5
numba==0.53.0
Pillow==8.1.2
scikit_learn==0.24.1
tensorflow==2.3.1
Keras
screeninfo